[libgens] VdpDebug.cpp: The test register is uint16_t, not uint8_t.
Not that it's being used for anything right now, but we should
use the correct data type anyway.

This fixes the following compiler warning on MSVC 2010:

src\libgens\Vdp\VdpDebug.cpp(144) : warning C4244: '=' : conversion from 'uint16_t' to 'uint8_t', possible loss of data

+++ src/libgens/Vdp/Vdp.hpp
@@ -246,8 +246,8 @@ class Vdp
 		int dbg_setAddress(uint32_t address);
 		int dbg_getCtrlLatch(int *latch) const;
 		int dbg_setCtrlLatch(int latch);
-		int dbg_getTestReg(uint8_t *out) const;
-		int dbg_setTestReg(uint8_t val);
+		int dbg_getTestReg(uint16_t *out) const;
+		int dbg_setTestReg(uint16_t val);
 
 		// TODO: Better VRAM writing functions.
 		int dbg_writeVRam_16(uint32_t address, const uint16_t *vram, int length);
